requirements.txt
accelerate==1.8.0
aiofiles==24.1.0       
aiohappyeyeballs==2.6.1
aiohttp==3.11.16       
aiosignal==1.3.2       
altair==5.5.0
annotated-types==0.7.0 
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
deep-translator==1.11.4
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
effdet==0.4.1
emoji==2.14.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langdetect==1.0.9
langsmith==0.3.30
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.43.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.19.2
openai==1.73.0
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==0.22.0
pikepdf==9.9.0
pillow==11.2.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pymongo==4.14.1
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdf==5.5.0
pypdfium2==4.30.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.46.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.4.0
uvicorn==0.34.1
watchdog==6.0.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.19.0
zstandard==0.23.0